refactor:extract callback functions
refactor:extract callback functions

refactor:extract randomize id function
